Moved ParseResult to types
